requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.7.2
openpyxl==3.1.2
jinja2==3.1.4
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.1
Pillow==10.3.0
python-dotenv==1.0.1
